Make new entry screen dynamic
