+++ app/protected/forms/page.tsx
@@ -28,7 +28,7 @@ import { format } from "date-fns";
 import { Sheet, SheetContent, SheetHeader, SheetTitle } from "@/components/ui/sheet";
 import TextareaAutosize from 'react-textarea-autosize';
 import { toast } from "sonner";
-import { Card } from "@/components/ui/card";
+import { Card, CardContent } from "@/components/ui/card";
 import {
   DndContext,
   closestCenter,
@@ -49,6 +49,10 @@ import { FormBuilderQuestion } from '@/lib/types/form-builder';
 import { AssigneeSelector } from '@/components/AssigneeSelector';
 import { useAuth } from '@/hooks/useAuth';
 import { getSupabaseClient } from '@/lib/supabase/client';
+import { createFormEntry } from "@/lib/api/form-entries";
+import { Textarea } from "@/components/ui/textarea";
+import { Label } from "@/components/ui/label";
+import { Checkbox } from "@/components/ui/checkbox";
 
 const questionTypes = [
   { value: "question", label: "Short answer" },
@@ -100,6 +104,14 @@ export default function FormsPage() {
   const [assignees, setAssignees] = useState<{ id: string; name: string; avatar_url?: string }[]>([]);
   const [assigneeError, setAssigneeError] = useState<string | null>(null);
 
+  // New state for entry creation
+  const [isCreatingEntry, setIsCreatingEntry] = useState(false);
+  const [entryName, setEntryName] = useState<string>("");
+  const [answers, setAnswers] = useState<Record<number, any>>({});
+  const [formErrors, setFormErrors] = useState<Record<number, string>>({});
+  const [isSubmittingEntry, setIsSubmittingEntry] = useState(false);
+  const [assignedBy, setAssignedBy] = useState<string | null>(null);
+
   const sensors = useSensors(
     useSensor(PointerSensor),
     useSensor(KeyboardSensor, {
@@ -107,13 +119,27 @@ export default function FormsPage() {
     })
   );
 
+  // Format current date for the entry name placeholder
+  const currentDate = new Intl.DateTimeFormat('en-US', {
+    month: 'short', 
+    day: 'numeric',
+    year: 'numeric'
+  }).format(new Date());
+
   // Effect to handle initial formId from URL
   useEffect(() => {
     const formId = searchParams.get('formId');
+    const entryMode = searchParams.get('entryMode');
+    
     if (formId) {
       const id = parseInt(formId);
       if (!isNaN(id)) {
         setSelectedFormId(id);
+        
+        // Set entry creation mode if requested
+        if (entryMode === 'new') {
+          setIsCreatingEntry(true);
+        }
       }
     }
   }, [searchParams]);
@@ -145,6 +171,7 @@ export default function FormsPage() {
       setEditedTitle("");
       setEditedItems([]);
       setIsEditing(false);
+      setIsCreatingEntry(false);
       return;
     }
 
@@ -155,6 +182,36 @@ export default function FormsPage() {
         setFormDetail(data);
         setEditedTitle(data?.form.name || "");
         setEditedItems(data?.items || []);
+
+        // Set default entry name with form name and date when form is loaded
+        if (data?.form) {
+          setEntryName(`${data.form.name} - ${currentDate}`);
+        }
+
+        // Determine who assigned the form
+        if (data?.form.owner_id === user?.id) {
+          setAssignedBy("Me");
+        } else if (data?.form.assignees && data.form.assignees.length > 0) {
+          // Use the first assignee's name if available
+          const assignee = data.form.assignees[0];
+          setAssignedBy(assignee.raw_user_meta_data.name || "Unknown");
+        } else if (data?.form.owner_id) {
+          // Fetch owner details from Supabase
+          const supabase = getSupabaseClient();
+          const { data: userData, error } = await supabase.rpc('get_user_details', {
+            user_ids: [data.form.owner_id]
+          });
+          
+          if (error) {
+            console.error('Error fetching owner details:', error);
+            setAssignedBy("Unknown");
+          } else if (userData && userData.length > 0) {
+            const ownerData = userData[0].raw_user_meta_data as { name?: string };
+            setAssignedBy(ownerData.name || "Unknown User");
+          } else {
+            setAssignedBy("Unknown");
+          }
+        }
 
         // Fetch assignees from entity_assignees
         const supabase = getSupabaseClient();
@@ -203,7 +260,7 @@ export default function FormsPage() {
     };
 
     fetchFormAndAssignees();
-  }, [selectedFormId]);
+  }, [selectedFormId, user?.id, currentDate]);
 
   const filteredForms = forms.filter(
     (form) => form.name.toLowerCase().includes(searchTerm.toLowerCase())
@@ -222,7 +279,11 @@ export default function FormsPage() {
 
   const handleCreateEntry = (formId: number | undefined) => {
     if (formId === undefined) return;
-    router.push(`/protected/forms/entries/new?formId=${formId}`);
+    setIsCreatingEntry(true);
+    setAnswers({});
+    setFormErrors({});
+    // Update URL to reflect entry creation mode
+    router.push(`/protected/forms?formId=${formId}&entryMode=new`);
   };
 
   const handleCloseDetail = () => {
@@ -230,7 +291,16 @@ export default function FormsPage() {
     setTimeout(() => {
       router.push('/protected/forms');
       setSelectedFormId(null);
+      setIsCreatingEntry(false);
     }, 300);
+  };
+
+  const handleCloseEntryForm = () => {
+    setIsCreatingEntry(false);
+    // Update URL to remove entry mode parameter but keep the form ID
+    if (selectedFormId) {
+      router.push(`/protected/forms?formId=${selectedFormId}`);
+    }
   };
 
   const handleSheetOpenChange = (isOpen: boolean) => {
@@ -463,6 +533,175 @@ export default function FormsPage() {
     }
   };
 
+  const handleAnswerChange = (itemId: number, value: any) => {
+    setAnswers((prev) => ({
+      ...prev,
+      [itemId]: value,
+    }));
+
+    // Clear error for this field if it exists
+    if (formErrors[itemId]) {
+      setFormErrors((prev) => {
+        const newErrors = { ...prev };
+        delete newErrors[itemId];
+        return newErrors;
+      });
+    }
+  };
+
+  const validateForm = () => {
+    const newErrors: Record<number, string> = {};
+    
+    // Check for entry name
+    if (!entryName.trim()) {
+      toast.error("Please provide an entry name");
+      return false;
+    }
+    
+    // Check for required fields
+    if (formDetail) {
+      formDetail.items.forEach((item) => {
+        if (item.is_required && (!answers[item.id!] || answers[item.id!] === "")) {
+          newErrors[item.id!] = "This field is required";
+        }
+      });
+    }
+    
+    setFormErrors(newErrors);
+    return Object.keys(newErrors).length === 0;
+  };
+
+  const handleSubmitEntry = async () => {
+    if (!user?.id || !formDetail || !selectedFormId) return;
+    
+    if (!validateForm()) {
+      toast.error("Please fill all required fields");
+      return;
+    }
+    
+    try {
+      setIsSubmittingEntry(true);
+      
+      // Prepare data for submission
+      const entryAnswers = Object.entries(answers).map(([itemId, value]) => ({
+        itemId: parseInt(itemId),
+        value: value
+      }));
+      
+      const response = await createFormEntry({
+        formId: selectedFormId,
+        userId: user.id,
+        name: entryName || formDetail.form.name,
+        answers: entryAnswers
+      });
+      
+      toast.success("Form entry submitted successfully");
+      
+      // Navigate to entries page and open the detail panel
+      router.push(`/protected/entries?entryId=${response.entry.id}`);
+    } catch (error) {
+      console.error("Error submitting form entry:", error);
+      toast.error("Failed to submit form entry");
+    } finally {
+      setIsSubmittingEntry(false);
+    }
+  };
+
+  const renderFormItem = (item: FormItem) => {
+    const itemId = item.id!;
+    const hasError = !!formErrors[itemId];
+    
+    switch (item.item_type) {
+      case "question":
+        return (
+          <div className="space-y-2">
+            <Label htmlFor={`question-${itemId}`} className="font-medium">
+              {item.question_value} {item.is_required && <span className="text-red-500">*</span>}
+            </Label>
+            <Input
+              id={`question-${itemId}`}
+              value={answers[itemId] || ""}
+              onChange={(e) => handleAnswerChange(itemId, e.target.value)}
+              className={hasError ? "border-red-500" : ""}
+            />
+            {hasError && <p className="text-sm text-red-500">{formErrors[itemId]}</p>}
+          </div>
+        );
+      
+      case "radio_box":
+        return (
+          <div className="space-y-2">
+            <Label className="font-medium">
+              {item.question_value} {item.is_required && <span className="text-red-500">*</span>}
+            </Label>
+            <Select
+              value={answers[itemId] || ""}
+              onValueChange={(value) => handleAnswerChange(itemId, value)}
+            >
+              <SelectTrigger className="w-full">
+                <SelectValue placeholder="Select option" />
+              </SelectTrigger>
+              <SelectContent>
+                {item.options?.filter(option => option.trim() !== "").map((option, index) => (
+                  <SelectItem key={index} value={option}>
+                    {option}
+                  </SelectItem>
+                ))}
+              </SelectContent>
+            </Select>
+            {hasError && <p className="text-sm text-red-500">{formErrors[itemId]}</p>}
+          </div>
+        );
+      
+      case "checklist":
+        return (
+          <div className="space-y-2">
+            <Label className="font-medium">
+              {item.question_value} {item.is_required && <span className="text-red-500">*</span>}
+            </Label>
+            <div className="space-y-2">
+              {item.options?.map((option, index) => (
+                <div key={index} className="flex items-center space-x-2">
+                  <Checkbox
+                    id={`checkbox-${itemId}-${index}`}
+                    checked={(answers[itemId] || []).includes(option)}
+                    onCheckedChange={(checked) => {
+                      const currentValues = answers[itemId] || [];
+                      const newValues = checked
+                        ? [...currentValues, option]
+                        : currentValues.filter((v: string) => v !== option);
+                      handleAnswerChange(itemId, newValues);
+                    }}
+                  />
+                  <Label htmlFor={`checkbox-${itemId}-${index}`}>{option}</Label>
+                </div>
+              ))}
+            </div>
+            {hasError && <p className="text-sm text-red-500">{formErrors[itemId]}</p>}
+          </div>
+        );
+      
+      case "photo":
+        // Placeholder for photo upload (would need additional components)
+        return (
+          <div className="space-y-2">
+            <Label className="font-medium">
+              {item.question_value} {item.is_required && <span className="text-red-500">*</span>}
+            </Label>
+            <Card className="bg-muted/40">
+              <CardContent className="flex flex-col items-center justify-center p-6">
+                <p className="text-muted-foreground text-sm">Photo upload not yet implemented</p>
+              </CardContent>
+            </Card>
+            {hasError && <p className="text-sm text-red-500">{formErrors[itemId]}</p>}
+          </div>
+        );
+      
+      default:
+        return null;
+    }
+  };
+
   return (
     <div className="container py-6 space-y-6">
       <div className="flex flex-col md:flex-row md:items-center md:justify-between gap-4">
@@ -615,6 +854,67 @@ export default function FormsPage() {
           ) : !formDetail ? (
             <div className="flex items-center justify-center h-full">
               <p>Form not found</p>
+            </div>
+          ) : isCreatingEntry ? (
+            // Form Entry Creation UI
+            <div className="h-full flex flex-col">
+              <div className="p-6 border-b">
+                <div className="flex items-start gap-4">
+                  <Button 
+                    variant="ghost"
+                    size="icon"
+                    onClick={handleCloseEntryForm}
+                  >
+                    <ArrowLeft className="h-4 w-4" />
+                  </Button>
+                  <div className="flex-1">
+                    <h2 className="text-2xl font-semibold mb-2">{formDetail.form.name}</h2>
+                    {assignedBy && (
+                      <p className="text-sm text-muted-foreground">Assigned by {assignedBy}</p>
+                    )}
+                  </div>
+                </div>
+              </div>
+              <div className="flex-1 overflow-auto">
+                <div className="p-6 space-y-6">
+                  <div className="space-y-4">
+                    <div className="space-y-2">
+                      <Label htmlFor="entry-name" className="font-medium">
+                        Entry Name <span className="text-red-500">*</span>
+                      </Label>
+                      <Input
+                        id="entry-name"
+                        value={entryName}
+                        onChange={(e) => setEntryName(e.target.value)}
+                        placeholder="Give this entry a name"
+                      />
+                    </div>
+                    
+                    {formDetail.items.map((item) => (
+                      <div key={item.id} className="border rounded-lg p-4">
+                        {renderFormItem(item)}
+                      </div>
+                    ))}
+                  </div>
+                </div>
+              </div>
+              <div className="p-6 border-t">
+                <div className="flex justify-end gap-2">
+                  <Button 
+                    variant="outline" 
+                    onClick={handleCloseEntryForm}
+                    disabled={isSubmittingEntry}
+                  >
+                    Cancel
+                  </Button>
+                  <Button 
+                    onClick={handleSubmitEntry}
+                    disabled={isSubmittingEntry}
+                  >
+                    {isSubmittingEntry ? "Submitting..." : "Submit"}
+                  </Button>
+                </div>
+              </div>
             </div>
           ) : (
             <div className="h-full flex flex-col">
